pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "podcaster"
version = "0.1.0"
description = "AI-powered podcast creation tool using LLMs and Speech2Text"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Podcaster Team", email = "contact@podcaster.example"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
]
llm = [
    "openai>=1.0.0",
    "anthropic>=0.7.0",
]
speech = [
    "openai-whisper>=20231117",
    "torch>=2.0.0",
]
tts = [
    "elevenlabs>=0.2.0",
    "edge-tts>=6.1.0",
    "TTS>=0.22.0",
]
audio = [
    "pydub>=0.25.0",
    "soundfile>=0.12.0",
    "ffmpeg-python>=0.2.0",
]
all = [
    "openai>=1.0.0",
    "anthropic>=0.7.0",
    "openai-whisper>=20231117",
    "torch>=2.0.0",
    "elevenlabs>=0.2.0",
    "edge-tts>=6.1.0",
    "TTS>=0.22.0",
    "pydub>=0.25.0",
    "soundfile>=0.12.0",
    "ffmpeg-python>=0.2.0",
]

[project.scripts]
podcaster = "scripts.cli:main"

[project.urls]
Homepage = "https://github.com/mikosovsky/podcaster"
Repository = "https://github.com/mikosovsky/podcaster"
Issues = "https://github.com/mikosovsky/podcaster/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/podcaster --cov-report=html --cov-report=term"
